service/requirements.txt
fastapi==0.95.2
uvicorn==0.23.1
prometheus_client==0.17.0
httpx==0.24.0
python-dotenv==1.0.0
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-instrumentation-fastapi==0.38b0
opentelemetry-exporter-jaeger==1.17.0
opentelemetry-instrumentation-requests
requests